Check request method and remove csrf_exempt
