+++ TrackApp/static/TrackApp/dashboard.js
@@ -1,3 +1,5 @@
+import * as utils from "./trackapp_utils.js";
+
 document.addEventListener('DOMContentLoaded', () => {
     // Load initial page
     let page = 1;
@@ -9,25 +11,6 @@ document.addEventListener('DOMContentLoaded', () => {
     }
 
 });
-
-
-function getCookie(name) {
-    /* Provided in the Django's documentation to get the csrf code
-    *  https://docs.djangoproject.com/en/4.0/ref/csrf/ */
-    let cookieValue = null;
-    if (document.cookie && document.cookie !== '') {
-        const cookies = document.cookie.split(';');
-        for (let i = 0; i < cookies.length; i++) {
-            const cookie = cookies[i].trim();
-            // Does this cookie string begin with the name we want?
-            if (cookie.substring(0, name.length + 1) === (name + '=')) {
-                cookieValue = decodeURIComponent(cookie.substring(name.length + 1));
-                break;
-            }
-        }
-    }
-    return cookieValue;
-}
 
 
 function page_management(page, number_pages) {
@@ -134,7 +117,7 @@ function remove_session(id, page) {
 
     // When click yes
     btn_yes.onclick = () => {
-        const csrftoken = getCookie('csrftoken');
+        const csrftoken = utils.getCookie('csrftoken');
         document.querySelector('#div_spinner_modal').style.display = 'inline-block';
         fetch(`/editor/remove_session/${id}`, {
             method: 'POST',
