+++ TrackApp/static/TrackApp/trackapp_utils.js
@@ -1,0 +1,17 @@
+export function getCookie(name) {
+    /* Provided in the Django's documentation to get the csrf code
+    *  https://docs.djangoproject.com/en/4.0/ref/csrf/ */
+    let cookieValue = null;
+    if (document.cookie && document.cookie !== '') {
+        const cookies = document.cookie.split(';');
+        for (let i = 0; i < cookies.length; i++) {
+            const cookie = cookies[i].trim();
+            // Does this cookie string begin with the name we want?
+            if (cookie.substring(0, name.length + 1) === (name + '=')) {
+                cookieValue = decodeURIComponent(cookie.substring(name.length + 1));
+                break;
+            }
+        }
+    }
+    return cookieValue;
+}
